+++ internal/backends/sqlite/backend.go
@@ -16,23 +16,17 @@ package sqlite
 
 import (
 	"context"
-	"fmt"
-	"net/url"
-	"os"
-	"strings"
 
 	"go.uber.org/zap"
 	_ "modernc.org/sqlite"
 
 	"github.com/FerretDB/FerretDB/internal/backends"
 	"github.com/FerretDB/FerretDB/internal/backends/sqlite/metadata"
-	"github.com/FerretDB/FerretDB/internal/util/lazyerrors"
 )
 
 // backend implements backends.Backend interface.
 type backend struct {
-	r   *metadata.Registry
-	uri *url.URL
+	r *metadata.Registry
 }
 
 // NewBackendParams represents the parameters of NewBackend function.
@@ -43,64 +37,14 @@ type NewBackendParams struct {
 
 // NewBackend creates a new SQLite backend.
 func NewBackend(params *NewBackendParams) (backends.Backend, error) {
-	uri, err := validateURI(params.URI)
-	if err != nil {
-		return nil, fmt.Errorf("failed to parse SQLite URI %q: %s", params.URI, err)
-	}
-
-	r, err := metadata.NewRegistry(uri, params.L.Named("metadata"))
-	if err != nil {
-		return nil, lazyerrors.Error(err)
-	}
-
-	return backends.BackendContract(&backend{
-		r:   r,
-		uri: uri,
-	}), nil
-}
-
-// validateURI checks given URI value and returns parsed URL.
-// URI should contain 'file' scheme and point to an existing directory.
-// Path should end with '/'. Authority should be empty or absent.
-//
-// Returned URL contains path in both Path and Opaque to make String() method work correctly.
-func validateURI(value string) (*url.URL, error) {
-	uri, err := url.Parse(value)
+	r, err := metadata.NewRegistry(params.URI, params.L.Named("metadata"))
 	if err != nil {
 		return nil, err
 	}
 
-	if uri.Scheme != "file" {
-		return nil, fmt.Errorf(`expected "file:" schema, got %q`, uri.Scheme)
-	}
-
-	if uri.User != nil {
-		return nil, fmt.Errorf(`expected empty user info, got %q`, uri.User)
-	}
-
-	if uri.Host != "" {
-		return nil, fmt.Errorf(`expected empty host, got %q`, uri.Host)
-	}
-
-	if uri.Path == "" && uri.Opaque != "" {
-		uri.Path = uri.Opaque
-	}
-	uri.Opaque = uri.Path
-
-	if !strings.HasSuffix(uri.Path, "/") {
-		return nil, fmt.Errorf(`expected path ending with "/", got %q`, uri.Host)
-	}
-
-	fi, err := os.Stat(uri.Path)
-	if err != nil {
-		return nil, fmt.Errorf(`%q should be an existing directory, got %s`, uri.Path, err)
-	}
-
-	if !fi.IsDir() {
-		return nil, fmt.Errorf(`%q should be an existing directory`, uri.Path)
-	}
-
-	return uri, nil
+	return backends.BackendContract(&backend{
+		r: r,
+	}), nil
 }
 
 // Close implements backends.Backend interface.
