+++ internal/backends/sqlite/metadata/registry.go
@@ -22,7 +22,6 @@ import (
 	"encoding/hex"
 	"errors"
 	"fmt"
-	"net/url"
 
 	"go.uber.org/zap"
 	"modernc.org/sqlite"
@@ -42,10 +41,10 @@ type Registry struct {
 }
 
 // NewRegistry creates a registry for the given directory.
-func NewRegistry(uri *url.URL, l *zap.Logger) (*Registry, error) {
-	p, err := pool.New(uri, l.Named("pool"))
+func NewRegistry(u string, l *zap.Logger) (*Registry, error) {
+	p, err := pool.New(u, l.Named("pool"))
 	if err != nil {
-		return nil, lazyerrors.Error(err)
+		return nil, err
 	}
 
 	return &Registry{
